Assignment for 09 Jul 2022 session
Problem Statement - A car company believes that percentage of residents in a city ABC that owns a vehicle is 60% or less. A sales manager disagrees with this. He conducts a hypothesis testing surveying 250 residents and found that 170 responded "Yes" to owning a vehicle.
a) State null and alternate hypothesis
b) At 10% significance value is there enough to support the idea that vehicle ownership in city ABC is 60% or less?