Statistics Assignment1
Que 1) Plot a histogram,
10, 13, 18, 22, 27, 32, 38, 40, 45, 51, 56, 57, 88, 90, 92, 94, 99

Que 2) In a quant test of the CAT Exam, the population standard deviation is known to be 100. A sample of 25 tests taken has a mean of 520. Construct an 80% CI about the mean.

Que 3) A car believes that the percentage of citizens in city ABC that owns a vehicle is 60% or less. A sales manager disagrees with this. He conducted a hypothesis testing surveying 250 residents & found that 170 residents responded yes to owning a vehicle.
a. State the null & alternate hypothesis.
b. At a 10% significance level, is there enough evidence to support the idea that vehicle owner in ABC city is 60% or less.

Que 4) What is the value of the 99 percentile?
2,2,3,4,5,5,5,6,7,8,8,8,8,8,9,9,10,11,11,12

Que 5) In left & right-skewed data, what is the relationship between mean, median & mode?
Draw the graph to represent the same.